+++ common/src/main/resources/assets/yama_copper_plates/block/copper_pressure_plate.png
