+++ common/src/main/resources/assets/yama_copper_plates/blockstates/copper_pressure_plate.json
