+++ common/src/main/resources/assets/yama_copper_plates/items/copper_pressure_plate.json
