+++ common/src/main/resources/assets/yama_copper_plates/lang/en_us.json
