+++ common/src/main/resources/assets/yama_copper_plates/models/block/copper_pressure_plate.json
