+++ common/src/main/resources/assets/yama_copper_plates/models/item/copper_pressure_plate.json
