+++ common/src/main/resources/assets/yama_copper_plates/textures/block/copper_pressure_plate.png
