+++ common/src/main/resources/data/yama_copper_plates/recipe/copper_pressure_plate.json
