+++ common/src/main/resources/icon.png
