--- fabric/src/main/resources/assets/yama_copper_plates/models/block/copper_pressure_plate_down.json
@@ -1,6 +1,0 @@
-{
-  "parent": "block/pressure_plate_down",
-  "textures": {
-    "texture": "yama_copper_plates:block/copper_pressure_plate"
-  }
-}
