+++ gradle.properties
@@ -15,6 +15,6 @@ yarn_mappings = 1.21.5+build.1
 # Dependencies
 architectury_api_version = 16.1.4
 fabric_loader_version = 0.16.14
-fabric_api_version = 0.124.2+1.21.5
-neoforge_version = 21.5.66-beta
+fabric_api_version = 0.126.0+1.21.5
+neoforge_version = 21.5.75
 yarn_mappings_patch_neoforge_version = 1.21+build.4
