--- neoforge/src/main/resources/assets/yama_copper_plates/blockstates/copper_pressure_plate.json
@@ -1,10 +1,0 @@
-{
-  "variants": {
-    "powered=false": {
-      "model": "yama_copper_plates:block/copper_pressure_plate"
-    },
-    "powered=true": {
-      "model": "yama_copper_plates:block/copper_pressure_plate_down"
-    }
-  }
-}
