--- neoforge/src/main/resources/assets/yama_copper_plates/lang/en_us.json
@@ -1,4 +1,0 @@
-{
-  "item.yama_copper_plates.copper_pressure_plate": "Waxed Lightly Weathered Pressure Plate of Copper (i)",
-  "block.yama_copper_plates.copper_pressure_plate": "Waxed Lightly Weathered Pressure Plate of Copper (b)"
-}
