--- neoforge/src/main/resources/assets/yama_copper_plates/models/item/copper_pressure_plate.json
@@ -1,3 +1,0 @@
-{
-  "parent": "yama_copper_plates:block/copper_pressure_plate"
-}
